pyproject.toml
[project]
name = "systems-explorer-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.1.40",
    "langchain-openai>=0.3.1",
    "playwright>=1.50.0",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
]
